requirements.txt
backports.entry-points-selectable==1.1.0
contourpy==1.0.5
cycler==0.11.0
distlib==0.3.3
filelock==3.0.12
fonttools==4.37.2
kiwisolver==1.4.4
matplotlib==3.6.0
numpy==1.21.4
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
platformdirs==2.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
virtualenv==20.8.1
wordcloud==1.8.2.2
